feat: toast on opérations (crud)

+++ client/src/components/DishesList/DishesList.tsx
@@ -5,6 +5,7 @@ import DisplayDishAccordion from "./DisplayDishAccordion";
 import { dishApi } from "@/services/api";
 import useDishes from "@/hooks/useDishes";
 import { useState } from "react";
+import { useToastStore } from "@/store";
 
 interface DishesListProps {
   restaurantId: number;
@@ -24,26 +25,49 @@ const DishesList: React.FC<DishesListProps> = ({
   const handleCancelAddOrUpdate = () => {
     setCreating(false);
   };
+
+  const setToast = useToastStore((state) => state.setToast);
   const handleCreateDish = (toCreate: CreateDishDTO) => {
     dishApi.create(restaurantId, toCreate).then(
       () => {
         setCreating(false);
         refreshDishes();
+        setToast({
+          message: "Plat créé avec succès",
+          type: "success",
+          isOpen: true,
+        });
       },
       (error) => {
         console.error("Erreur lors de la création du plat:", error);
+        setToast({
+          message: "Erreur lors de la création du plat",
+          type: "error",
+          isOpen: true,
+        });
       }
     );
   };
   const handleDeleteDish = (dishId: number) => () => {
-    dishApi.delete(restaurantId, dishId).then(
-      () => {
-        refreshDishes();
-      },
-      (error) => {
-        console.error("Erreur lors de la suppression du plat:", error);
-      }
-    );
+    window.confirm("Voulez-vous vraiment supprimer ce plat ?") &&
+      dishApi.delete(restaurantId, dishId).then(
+        () => {
+          refreshDishes();
+          setToast({
+            message: "Plat supprimé avec succès",
+            type: "success",
+            isOpen: true,
+          });
+        },
+        (error) => {
+          console.error("Erreur lors de la suppression du plat:", error);
+          setToast({
+            message: "Erreur lors de la suppression du plat",
+            type: "error",
+            isOpen: true,
+          });
+        }
+      );
   };
 
   if (loading) return <Box>Chargement des plats...</Box>;
